Refactor : [WUMO-326] 후보지 삭제는 작성자 및 모임장 할 수 있다.

+++ src/main/java/org/prgrms/wumo/domain/location/service/LocationService.java
@@ -87,6 +87,8 @@ public class LocationService {
 	public void deleteLocation(Long locationId){
 		Location location = getLocationEntity(locationId);
 
+		checkAuthorization(location, getMemberId());
+
 		locationRepository.deleteById(locationId);
 	}
